// PortalApp/ClientApp/src/enums/event.enum.ts
export enum Category {
    LANDINGPAGE = 'LandingPage',
}
export enum Action {
    CREATE = 'Create',

    DELETE = 'Delete',

    UPDATE = 'Update',

    DUPLICATE = 'Duplicate',

    SAVE = 'Save',

    CANCEL = 'Cancel',

    NEXTRECORD = 'Next Record',

    SEARCH = 'Search',

    VIEWEDITHISTORYlIST = 'View Edit Version History List',

    VIEWDOWNLOADHISTORYLIST = 'View Download Version History List',

    VIEWLAUNCHHISTORYLIST = 'View Launch Version History List',

    OPENREPORTTOEDIT = 'Open Report To Edit',

    DOWNLOADREPORTTOVIEW = 'Download Report To View',

    LAUNCHREPORTTOVIEW = 'Launch Report To View',
}

export enum Module {
    FLOWCHART = 'Flowchart',

    EDIT = 'Edit',

    DOWNLOAD = 'Download',

    LAUNCH = 'Report',
}

export enum SubModule {
    HISTORY = 'History',
}